修改力矩曲线载重单位，修改实时数据力矩

+++ app/src/main/java/com/ztrs/zgj/device/bean/RealTimeDataBean.java
@@ -109,7 +109,7 @@ public class RealTimeDataBean {
     int xSlope;
     int ySlope;
 
-    short torque;
+    int torque;
     byte windLevel;
 
     byte wireRopeState;
@@ -880,11 +880,11 @@ public class RealTimeDataBean {
         this.ySlope = ySlope;
     }
 
-    public short getTorque() {
+    public int getTorque() {
         return torque;
     }
 
-    public void setTorque(short torque) {
+    public void setTorque(int torque) {
         this.torque = torque;
     }
